Move methods from PlanVariableAllocator to VariableAllocator
Currently, both PlanVariableAllocator and VariableAllocator are
responsible for similar tasks, leading to redundancy in the codebase.
To address this, we plan to remove PlanVariableAllocator and
consolidate all functionality within VariableAllocator.

This change moves many methods from PlanVariableAllocator to
VariableAllocator.
